Removed GridList Component. Entirely redundant
